backend/requirements.txt
# Core API framework
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.20

# OCR and Image Processing
numpy>=2.0.0
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.0.0

# Claude AI
anthropic>=0.40.0
python-dotenv>=1.0.0

# Browser automation and Langchain
langchain-anthropic>=0.3.0
browser-use>=0.1.0

# Additional utilities
requests>=2.31.0
